Fixed bug about output notice message when clear cache without cache file.
git-svn-id: file://localhost/home/repos/svn/wizin/trunk@324 fd31610b-4be5-dc11-a523-00018064d71f

+++ wizin/src/cache/File.class.php
@@ -76,7 +76,9 @@ if (! class_exists('Wizin_Cache')) {
 
         function clear()
         {
-            unlink($this->_sCacheFile);
+            if (file_exists($this->_sCacheFile)) {
+                unlink($this->_sCacheFile);
+            }
         }
 
     }
